IOW-307 Moved a legend test per Mary and updated Changelog.
